+++ packages/plugin/ai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dovenv/ai
 
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @dovenv/core@0.0.14
+
 ## 0.0.13
 
 ### Patch Changes